refactor: change product data table  into product specifications page and change the old one to product spec archive page and hide it from the sidebar.
